new version login and home changes

+++ src/app/components/home/home.component.ts
@@ -1,5 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Inject, OnInit } from '@angular/core';
+import { MAT_DIALOG_DATA, MatDialog, MatDialogRef } from '@angular/material/dialog';
 import { Router } from '@angular/router';
+import { MaterialModule } from 'src/app/modules/material/material.module';
 import { ApiService } from 'src/app/services/api.service';
 import { AuthService } from 'src/app/services/auth.service';
 
@@ -12,7 +14,8 @@ export class HomeComponent implements OnInit {
   constructor(
     private api: ApiService,
     private router: Router,
-    public authService: AuthService
+    public authService: AuthService,
+    public dialog: MatDialog
   ) {
     window.onbeforeunload = (event) => {
       this.authService.logout();
@@ -27,11 +30,33 @@ export class HomeComponent implements OnInit {
     );
   }
   deleteBook(id: any): void {
-    this.api.deleteBook(`/books/${id}`).subscribe((data: any) => {
-      var newBooks = this.books.filter((book: any) => {
-        return book._id != id;
-      });
-      this.books = newBooks;
+    const dialogRef=this.dialog.open(DeleteDialogComponent, {
+      data:true,
+      width: '250px',
     });
+    dialogRef.afterClosed().subscribe((result)=>{
+      if(result){
+        this.api.deleteBook(`/books/${id}`).subscribe((data: any) => {
+          var newBooks = this.books.filter((book: any) => {
+            return book._id != id;
+          });
+          this.books = newBooks;
+        });
+      }
+    })
+  }
+}
+@Component({
+  selector: 'delete-dialog',
+  template: '<h1 mat-dialog-title>Delete Book</h1><div mat-dialog-content>Would you like to delete this book?</div><div mat-dialog-actions><button mat-button mat-dialog-close>No</button><button mat-button [mat-dialog-close]="data" cdkFocusInitial>Yes</button></div>',
+  styles: [],
+  standalone: true,
+  imports: [MaterialModule]
+})
+export class DeleteDialogComponent {
+
+  constructor(public dialogRef: MatDialogRef<DeleteDialogComponent>, @Inject(MAT_DIALOG_DATA) public data: any) { }
+  public confirmDelete(): void {
+
   }
 }
